refactor: OpenMajor 화면 -> OpenMajorBottomSheet 로 변경
- OpenMajor 화면 제거
- Timetable - OpenMajor BottomSheet 추가
- BottomSheet Composable isSheetOpen 파라미터 제거
