+++ composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/presentation/timetable/openlecture/OpenLectureContract.kt
@@ -18,6 +18,7 @@ data class OpenLectureState(
     val openLectureList: PersistentList<OpenLecture> = persistentListOf(),
     val selectedOpenMajor: String = "전체",
     val showSchoolLevelBottomSheet: Boolean = false,
+    val showOpenMajorBottomSheet: Boolean = false,
     val schoolLevel: SchoolLevel = SchoolLevel.ALL,
     val showSelectCellColorBottomSheet: Boolean = false,
     val selectedTimetableCellColor: TimetableCellColor = TimetableCellColor.BROWN,
@@ -51,7 +52,6 @@ fun TimetableDay.toText() = when (this) {
 sealed interface OpenLectureSideEffect {
     data object ScrollToTop : OpenLectureSideEffect
     data object PopBackStack : OpenLectureSideEffect
-    data class NavigateOpenMajor(val selectedOpenMajor: String) : OpenLectureSideEffect
     data class NavigateCellEditor(val argument: CellEditorArgument) : OpenLectureSideEffect
     data object NavigateAddCustomTimetableCell : OpenLectureSideEffect
     data class HandleException(val throwable: Throwable) : OpenLectureSideEffect
