+++ composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/presentation/timetable/timetable/component/bottomsheet/openmajor/OpenMajorBottomSheetContract.kt
@@ -1,6 +1,6 @@
-package com.chukchukhaksa.mobile.presentation.openmajor
+package com.chukchukhaksa.mobile.presentation.timetable.timetable.component.bottomsheet.openmajor
 
-import com.chukchukhaksa.mobile.presentation.openmajor.model.OpenMajor
+import com.chukchukhaksa.mobile.presentation.timetable.timetable.component.bottomsheet.openmajor.model.OpenMajor
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 
@@ -20,6 +20,4 @@ sealed interface OpenMajorSideEffect {
     data class HandleException(val throwable: Throwable) : OpenMajorSideEffect
     data object PopBackStack : OpenMajorSideEffect
     data class PopBackStackWithArgument(val argument: String) : OpenMajorSideEffect
-
-    data object ShowNeedLoginToast : OpenMajorSideEffect
 }
